Show publisher name on wide and narrow window headers (#709)
